% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyjWidget.R
\name{cyjWidget}
\alias{cyjWidget}
\alias{printf}
\title{cyjsWidget}
\usage{
printf(...)
}
\value{
shiny code which, wnen invoked (almost always by the shinyApp function, returns html

nothing

nothing
}
\description{
an R6 class providing clean access to the cyjShiny package
}
\details{
R6 Class representing an igv.js genome browser


This class provides the cytoscape.js graph library in a classic object-oriented-programming R context
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{cyjWidget$new()}}
\item \href{#method-ui}{\code{cyjWidget$ui()}}
\item \href{#method-server}{\code{cyjWidget$server()}}
\item \href{#method-fitGraph}{\code{cyjWidget$fitGraph()}}
\item \href{#method-clone}{\code{cyjWidget$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new cyjWidget
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cyjWidget$new(id, graph.json, width = "95\%", height = "500px", border = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{the html document div id}

\item{\code{graph.json}}{a json graph}

\item{\code{width}}{integer, in width}

\item{\code{height}}{integer, in pixels}

\item{\code{border}}{character string, for instance "1px solid purple; border-radius: 5px;"}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `cyjWidget` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ui"></a>}}
\if{latex}{\out{\hypertarget{method-ui}{}}}
\subsection{Method \code{ui()}}{
defines the html structure of this widget
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cyjWidget$ui()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-server"></a>}}
\if{latex}{\out{\hypertarget{method-server}{}}}
\subsection{Method \code{server()}}{
defines the event handlers used with the ui
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cyjWidget$server(input, output, session)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{list, managed by Shiny}

\item{\code{output}}{list, managed by Shiny}

\item{\code{session}}{list (environment?) managed by Shiny}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fitGraph"></a>}}
\if{latex}{\out{\hypertarget{method-fitGraph}{}}}
\subsection{Method \code{fitGraph()}}{
scales the entire graph to fit symetrically in its div
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cyjWidget$fitGraph()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cyjWidget$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
